backend/requirements.txt
fastapi
uvicorn
duckduckgo-search
python-dotenv
aiohttp
